Created Project
So this is just my endeavor in figuring out the geometry of 3D
graphics. I want to be able to define points in 3D space and be able to
view them on a 2D screen as if I was in that space.

Here I created the folders and stuff. My first try doing this I had
some trouble with the Github desktop application, so I figure if I
start it online I won’t run into as many problems there.