Added previously worked on stuff
I figured out as much as I could on paper, these are the beginnings of
my code. The main thing is separated into two packages: graphics and
viewer.

graphics is the graphics engine or whatever, both the actual space and
the objects inside it (though this may change) and the rules that
translate a view of that 3D space to a 2D image, for now made of points
and lines.

viewer will contain the java-specific stuff that translates that
theoretical image into something that can be drawn on the screen.

The idea is that the stuff in graphics can be translated into other
languages and other rendering frameworks.

Also gabrielsUtilities just has some useful stuff I wrote—array
manipulation, sorting, rounding, that sort of thing.