First draft of 3D to 2D + Maintenance
View will be the locations of points on the 2D screen. I’m thinking of
moving some of this to a different folder for geometry and just having
drawing stuff (connecting points, etc.) in graphics. We’ll see.

getVertPos and getHoriPos are given a point in space and both return
the distance of the point from the bottom or right of the View,
respectively. They’re just proxy, convenience methods, passing stuff
onto getPos.

getPos looks at where the viewpoint is, where the observed point is,
and what the field of view is, and figures out the distance of the
point to the edge of that plane of the field of view, then returns that
distance proportionate to the size of the screen. Imagine that there’s
a square pyramid, and the tip of the pyramid is the viewpoint. That
pyramid is your full range of view. Anything within the pyramid can be
seen, and the size of the thing in your field of vision is equal to the
proportion of the actual size of the thing to the are of the bottom of
the pyramid if you imagine the pyramid stops at that thing. The pyramid
extends infinitely, so the bottom will be huge at far-away objects, so
everything will appear much smaller. That’s the principle here,
expressed trigonometrically. getPos doesn’t imagine the full plane of
the pyramid, though—it only operates one-dimensionally. You give it x
and z coords and it’ll give you horizontal distance, you give it x and
y coords and it’ll give you vertical distance. That’s what the other
methods do.

Also done here: changed Graphics to View, because I feel it’s more
fitting.