Changed structure, added limits to view angle
I changed the structure here so Perspective is a general class for
holding methods for this sort of thing and View adds all the specific
parameters and is in general a more useful, substantiatable object.

Added 2 methods to Perspective: checkViewAngleContinuous and
checkViewAngleLimited.

The first returns a rolled over angle if it’s below or above certain
values, so I don’t get angles of like 12π

The second returns the angle reduced to the limits if it’s beyond those
limits.

Both are only for one angle, ‘cos I wanted something that could
actually return and it’d be more efficient to run through all that if
statement stuff once and change the second angle the same amount as the
first.