Movement of View, limitations identified
View:
added an associated Space to the View
added, then discarded, rotation on the yz-plane
identified possible limitation: no idea what will happen to points
outside of the viewing angles; this is a problem because to draw
polygons, the lines and shapes need to still be visible even after the
vertices are outside of the field of view. The math may take care of
it, but I’m not sure.

Space:
added array of Points and associated functionality so that it actually
functions as the container it’s meant to be

Point:
added array manipulation stuff so that I can mess with Point arrays
without casting up and down to Objects and back to Points
added setX, setY, and setZ methods for changing the location of the
point

Perspective:
changed the name of a variable for more clarity

Person:
this is a new class, extended from View, that adds methods to the View
to change location and angle of the View. Right now it’s just velocity,
no acceleration. Mouse motion I think will be the most difficult here;
I’ma try to keep the logic for it outside the applet, but
MouseMotionListener might be too specific for the logic to be useful at
all. We’ll see what happens.